+++ src/main/java/uk/gov/hmcts/reform/finrem/caseorchestration/service/stoprepresentingclient/LitigantRevocation.java
@@ -1,0 +1,8 @@
+package uk.gov.hmcts.reform.finrem.caseorchestration.service.stoprepresentingclient;
+
+public record LitigantRevocation(boolean applicantSolicitorRevoked, boolean respondentSolicitorRevoked) {
+
+    public boolean wasRevoked() {
+        return applicantSolicitorRevoked || respondentSolicitorRevoked;
+    }
+}
